+++ features/Context/WebUser.php
@@ -1420,54 +1420,6 @@ class WebUser extends RawMinkContext
     }
 
     /**
-     * @param string $products
-     *
-     * @When /^I mass-edit products (.*)$/
-     */
-    public function iMassEditProducts($products)
-    {
-        $page = $this->getPage('Product index');
-
-        foreach ($this->listToArray($products) as $product) {
-            $page->selectRow($product);
-        }
-
-        $page->massEdit();
-        $this->wait();
-    }
-
-    /**
-     * @param string $products
-     *
-     * @Then /^I select products (.*)$/
-     */
-    public function iSelectProducts($products)
-    {
-        $page = $this->getPage('Product index');
-
-        foreach ($this->listToArray($products) as $product) {
-            $page->selectRow($product);
-        }
-    }
-
-    /**
-     * @param string $products
-     *
-     * @When /^I mass-delete products (.*)$/
-     */
-    public function iMassDeleteProducts($products)
-    {
-        $page = $this->getPage('Product index');
-
-        foreach ($this->listToArray($products) as $product) {
-            $page->selectRow($product);
-        }
-
-        $page->massDelete();
-        $this->wait();
-    }
-
-    /**
      * @param string $operation
      *
      * @Given /^I choose the "([^"]*)" operation$/