+++ src/Pim/Bundle/CatalogBundle/DependencyInjection/PimCatalogExtension.php
@@ -32,6 +32,7 @@ class PimCatalogExtension extends Extension
         $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('parameters.yml');
         $loader->load('services.yml');
+        $loader->load('event_listeners.yml');
         $loader->load('managers.yml');
         $loader->load('attribute_types.yml');
         $loader->load('attribute_constraint_guessers.yml');