+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - MongoDB support (WIP)
 - Allow to add many quick export on datagrids
 - Add a parameter to ProductManager::save() and ProductManager::saveAll() to allow saving products without completeness recalculation
+- Optimize products mass deletion
 
 ## Bug fixes
 - Fixed verbose option always used
@@ -46,6 +47,8 @@
 - Move `DataGridBundle/Extension/Formatter/Property/*` to `DataGridBundle\Extension\Formatter\Property\ProductValue\*`
 - Move mass export in specific controller
 - Add an affectsCompleteness method to MassEditActionInterface to indicate whether performing the mass action requires recalculating the product completeness
+- Remove DeleteMassActionHandler, replaced by ProductDeleteMassActionHandler
+- Add deleteProductIds in ProductRepository
 
 # 1.0.2
 ## Bug Fixes